Merge /wiki/Article with /wiki/full/Article
To keep consistency with wiki urls.

Sorry about the big commit.

Now if you visit /wiki/Article and you weren't there before it'll show the
summary version. You can expand it and you will get the longer version.

As long as the longer version is available in memory that is what you will see.

When we add caching, we'll cache across sessions the summary & content
& expanded so that if you return to an article you can see it as you were
seeing it before.

For now, this is sufficient.

* NOTE: Full article is supported with a query parameter so that you can go to
it in the HTML only version