Fix scroll jank on mobile
Nesting the content inside of the menu and setting overflow auto on the .App
was causing huge jank on mobile chrome when scrolling through long articles and
the lazy loading the images.

Solution is avoid nesting the app content inside the menu container, and
instead keep the menu layers and the content layers (App-viewport) side by side
inside App.

This fixes the scroll jank completely (it took me a lot of debugging and
trying...).

I could only find this bug as a confirmation that I wasn't going crazy:
https://code.google.com/p/chromium/issues/detail?id=540604