Pull out webpack compilation from dev server
Run the compiler in parallel with webpack for faster restart times on the node
server and even faster reloads

Before it was frequent to try to reload but the node server would be stuck
compiling assets and the reload would take longer or fail. Having it this way
I haven't experienced that slowness, seems to be a lot faster now.